+++ frontend/src/models/levels/export-url.tsx
@@ -1,6 +1,11 @@
 import { createEffect, createEvent, merge, sample } from "effector";
 import { TEST_DEMO_URL, TEST_LEVEL_URL } from "configs";
-import { getDriverFormat, IBaseLevel, levelSupportsDemo } from "drivers";
+import {
+  getDriver,
+  getDriverFormat,
+  IBaseLevel,
+  levelSupportsDemo,
+} from "drivers";
 import { Trans } from "i18n/Trans";
 import { ColorType } from "ui/types";
 import { base64Encode } from "utils/encoding/base64";
@@ -22,11 +27,13 @@ export const exportLevelAsLink = async (
   }
 
   const driverName = $currentDriverName.getState()!;
+  const { applyLocalOptions } = getDriver(driverName)!;
   const { writeLevelset, createLevelset } = getDriverFormat(
     driverName,
     $currentDriverFormat.getState()!,
   )!;
   const url = new URL(baseUrl);
+  applyLocalOptions?.(level, url);
   const raw = writeLevelset(createLevelset([level]));
   const compressed = await tryGzipCompress(raw);
   url.hash = compressed ? "gz," + base64Encode(compressed) : base64Encode(raw);