+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Fix: Some UI texts correction in progress. Thanks to Lana Snik-Snak!
 - Fix: H/V/X Ports when set as Special wasn't marked Special in "Copy as Image".
 - Fix: Copy/Export Level/Selection icons and compact layout.
+- Fix: Copy level as test/demo link missed custom Options.
 - Enh: H/V/X Ports in Toolbar are halved too.
 - Enh: Spec Ports DB dialog UX:
   - display the tile in the giver cell;